use floats everywhere & disable minisolver

+++ src/tpn_table.cpp
@@ -3,8 +3,8 @@
 bool TpnTable::lookup_value(int max_depth, Value &value) const {
   bool found = lookup_value_normed(max_depth, value);
 
-  value.lower_bound += game_.tricks_taken_by_ns();
-  value.upper_bound += game_.tricks_taken_by_ns();
+  value.lower_bound += (float)game_.tricks_taken_by_ns();
+  value.upper_bound += (float)game_.tricks_taken_by_ns();
   if (value.pv_play.has_value()) {
     value.pv_play = game_.denormalize_card(*value.pv_play);
   }
@@ -14,8 +14,8 @@ bool TpnTable::lookup_value(int max_depth, Value &value) const {
 
 void TpnTable::upsert_value(int max_depth, const Value &value) {
   Value normed_value = {
-      .lower_bound = value.lower_bound - game_.tricks_taken_by_ns(),
-      .upper_bound = value.upper_bound - game_.tricks_taken_by_ns(),
+      .lower_bound = value.lower_bound - (float)game_.tricks_taken_by_ns(),
+      .upper_bound = value.upper_bound - (float)game_.tricks_taken_by_ns(),
   };
   if (value.pv_play.has_value()) {
     normed_value.pv_play = game_.normalize_card(*value.pv_play);
@@ -34,7 +34,7 @@ bool TpnTable::lookup_value_normed(int max_depth, Value &value) const {
     return true;
   } else {
     value.lower_bound = 0;
-    value.upper_bound = game_.tricks_left();
+    value.upper_bound = (float)game_.tricks_left();
     value.pv_play     = std::nullopt;
     return false;
   }
@@ -45,8 +45,8 @@ void TpnTable::upsert_value_normed(int max_depth, const Value &value) {
   auto  it    = table_.find(state);
   if (it != table_.end()) {
     Entry &entry      = it->second;
-    entry.lower_bound = (int8_t)value.lower_bound;
-    entry.upper_bound = (int8_t)value.upper_bound;
+    entry.lower_bound = value.lower_bound;
+    entry.upper_bound = value.upper_bound;
     entry.max_depth   = (int8_t)max_depth;
     entry.pv_play     = value.pv_play;
   } else {
@@ -54,8 +54,8 @@ void TpnTable::upsert_value_normed(int max_depth, const Value &value) {
         std::piecewise_construct,
         std::forward_as_tuple(state),
         std::forward_as_tuple(
-            (int8_t)value.lower_bound,
-            (int8_t)value.upper_bound,
+            value.lower_bound,
+            value.upper_bound,
             (int8_t)max_depth,
             value.pv_play
         )
